Move orca grid references into local orca grid
Try to remove orca_grid references from meshgenerator. Still a little
work to do with the topology and normalisation components, but once
finished we ought to be able to get the mesh generator to only fill the
mesh based on the local_orca_grid and only indexing using the i,j
indices for the local_orca_grid. Inside that object we handle the
conversions between global indices and rectangle indices.